Indicate OpenPGP certificate validity
(no functional changes here, just a variable change)

in /usr/share/doc/gnupg/DETAILS.gz, we see that TRUST_* has to do with
the validty of the OpenPGP certificate (which GnuPG ambiguously calls
the "key"):

-----
*** TRUST_
    These are several similar status codes:

    - TRUST_UNDEFINED <error_token>
    - TRUST_NEVER     <error_token>
    - TRUST_MARGINAL  [0  [<validation_model>]]
    - TRUST_FULLY     [0  [<validation_model>]]
    - TRUST_ULTIMATE  [0  [<validation_model>]]

    For good signatures one of these status lines are emitted to
    indicate the validity of the key used to create the signature.
    The error token values are currently only emitted by gpgsm.

    VALIDATION_MODEL describes the algorithm used to check the
    validity of the key.  The defaults are the standard Web of Trust
    model for gpg and the standard X.509 model for gpgsm.  The
    defined values are

       - pgp   :: The standard PGP WoT.
       - shell :: The standard X.509 model.
       - chain :: The chain model.
       - steed :: The STEED model.
       - tofu  :: The TOFU model

    Note that the term =TRUST_= in the status names is used for
    historic reasons; we now speak of validity.
-----

I believe that this corresponds to the highest validity level of all
OpenPGP User IDs in the corresponding certificate.

It's not clear to me that requiring a fully-valid User ID on the
signing key *is* necessarily the right semantics for what gpgverify is
trying to do, but the code should be clear about what it is doing.